# server/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PTCF Room & Equipment Reservation System API",
    "description": "API documentation for the UPLB ICropS Plant Tissue Culture Facility reservation management system.",
    "version": "1.0.0",
    "contact": {
      "name": "PTCF Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Local development server"
    },
    {
      "url": "https://your-render-url.onrender.com/api",
      "description": "Production server (Render)"
    }
  ],
...
  "paths": {
...
    "/equipment": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "summary": "Get all equipment",
        "description": "Returns a list of all equipment items. Accessible to all authenticated users.",
        "operationId": "getAllEquipment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of equipment retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Equipment"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "post": {
        "tags": [
          "Equipment"
        ],
        "summary": "Create new equipment",
        "description": "Creates a new equipment item with optional image upload. Requires staff or admin role.",
        "operationId": "createEquipment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "category",
                  "description"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Equipment name",
                    "example": "Laminar Flow Hood"
                  },
                  "category": {
                    "type": "string",
                    "description": "Equipment category",
                    "example": "Sterilization Equipment"
                  },
                  "description": {
                    "type": "string",
                    "description": "Detailed description",
                    "example": "Class II Biological Safety Cabinet"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "available",
                      "unavailable",
                      "maintenance"
                    ],
                    "description": "Availability status (optional, defaults to 'available')",
                    "example": "available"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file (optional, max 5MB, JPEG/PNG/GIF/WebP)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Equipment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipment"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Missing required fields: name, category, and description are required"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "description": "Server error - Failed to upload image or create equipment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/equipment/{id}": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "summary": "Get equipment by ID",
        "description": "Returns a single equipment item by ID. Accessible to all authenticated users.",
        "operationId": "getEquipmentById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Equipment ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Equipment retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipment"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "Equipment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Equipment not found"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Equipment"
        ],
        "summary": "Update equipment",
        "description": "Updates an existing equipment item. Can optionally update the image. Requires staff or admin role.",
        "operationId": "updateEquipment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Equipment ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Equipment name (optional)"
                  },
                  "category": {
                    "type": "string",
                    "description": "Equipment category (optional)"
                  },
                  "description": {
                    "type": "string",
                    "description": "Detailed description (optional)"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "available",
                      "unavailable",
                      "maintenance"
                    ],
                    "description": "Availability status (optional)"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "New image file (optional, replaces existing image)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Equipment updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipment"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "description": "Equipment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Equipment"
        ],
        "summary": "Delete equipment",
        "description": "Deletes an equipment item by ID. Requires staff or admin role.",
        "operationId": "deleteEquipment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Equipment ID",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Equipment deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Equipment deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "description": "Equipment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid credentials"
          }
        }
      },
...
      "Equipment": {
        "type": "object",
        "description": "Equipment model schema (for future Milestone 3)",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "description": "Equipment name"
          },
          "category": {
            "type": "string",
            "description": "Equipment category"
          },
          "description": {
            "type": "string",
            "description": "Detailed description"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri",
            "description": "Cloudinary image URL"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "unavailable",
              "maintenance"
            ],
            "description": "Availability status"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Unauthorized - Missing or invalid JWT token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "missingToken": {
                "value": {
                  "message": "Missing Bearer token"
                }
              },
              "invalidToken": {
                "value": {
                  "message": "Invalid or expired token"
                }
              }
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Forbidden - Insufficient role (requires staff or admin)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "Forbidden: insufficient role"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter ONLY your JWT token (without 'Bearer' prefix). Example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      }
    }
  },
  "security": []
}